requirements.txt
streamlit==1.33.0
openai>=1.12.0
python-dotenv==1.0.1
langchain==0.1.13
langchain-openai==0.1.0
langchain-community==0.0.29
chromadb==0.4.24
pypdf2==3.0.1
tiktoken==0.6.0 